refactor: user entity 경로 변경
