+++ src/modules/comments/comments-impl.service.ts
@@ -11,7 +11,7 @@ import { UsersRepository } from '../users/users.repository';
 import { PostsRepository } from '../posts/posts.repository';
 import { CommentsService } from './comment.service';
 import { Post } from '../posts/entities/post.entity';
-import { User } from '../users/user.entity';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class CommentsServiceImpl implements CommentsService {
