+++ src/modules/comments/entities/comment.entity.ts
@@ -8,7 +8,7 @@ import {
   ManyToOne,
 } from 'typeorm';
 import { BaseModel } from '../../../config/typeorm/base.entity';
-import { User } from '../../users/user.entity';
+import { User } from '../../users/entities/user.entity';
 import { Post } from '../../posts/entities/post.entity';
 
 @Entity('comment')
