+++ src/modules/posts/posts-impl.service.ts
@@ -6,7 +6,7 @@ import {
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
-import { User } from '../users/user.entity';
+import { User } from '../users/entities/user.entity';
 import { Post } from './entities/post.entity';
 import { PostsService } from './posts.service';
 import { Category } from '../categorys/entities/category.entity';
