+++ src/modules/posts/posts.module.ts
@@ -4,7 +4,7 @@ import { PostsController } from './posts.controller';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
 import { PostsRepository } from './posts.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../users/user.entity';
+import { User } from '../users/entities/user.entity';
 import { PostsService } from './posts.service';
 import { Category } from '../categorys/entities/category.entity';
 import { UsersRepository } from '../users/users.repository';
