+++ src/modules/users/users.controller.spec.ts
@@ -5,7 +5,7 @@ import { UsersRepository } from './users.repository';
 import { AuthService } from '../../auth/auth.service';
 import { JwtService } from '@nestjs/jwt';
 import { CreateUserDto } from './dto/create-user.dto';
-import { User } from './user.entity';
+import { User } from './entities/user.entity';
 import { UserInfo } from '../../auth/jwt.strategy';
 import { ConfigService } from '@nestjs/config';
 import { Response } from 'express';
