+++ src/modules/users/users.controller.ts
@@ -9,7 +9,7 @@ import {
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
-import { User } from './user.entity';
+import { User } from './entities/user.entity';
 import { AuthService } from '../../auth/auth.service';
 import { LocalAuthGuard } from '../../auth/local-auth.guard';
 import { UserInfo } from '../../auth/jwt.strategy';
