+++ src/test/repository/comments.repository.spec.ts
@@ -3,7 +3,7 @@ import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module'
 import { CommentsRepository } from '../../modules/comments/comments.repository';
 import { UsersRepository } from '../../modules/users/users.repository';
 import { TestTypeOrmModule } from '../db/test-db.module';
-import { User } from '../../modules/users/user.entity';
+import { User } from '../../modules/users/entities/user.entity';
 import { Comment } from '../../modules/comments/entities/comment.entity';
 import { PostsRepository } from '../../modules/posts/posts.repository';
 import { Post } from '../../modules/posts/entities/post.entity';
