+++ src/test/service/users.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { UsersService } from '../../modules/users/users.service';
 import { UsersRepository } from '../../modules/users/users.repository';
 import { CreateUserDto } from '../../modules/users/dto/create-user.dto';
-import { User } from '../../modules/users/user.entity';
+import { User } from '../../modules/users/entities/user.entity';
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { TestTypeOrmModule } from '../db/test-db.module';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
